+++ sample/coroutines/shared/src/commonMain/kotlin/com/arkivanov/mvikotlin/sample/coroutines/shared/details/Mappers.kt
@@ -2,8 +2,8 @@ package com.arkivanov.mvikotlin.sample.coroutines.shared.details
 
 import com.arkivanov.mvikotlin.sample.coroutines.shared.details.DetailsView.Event
 import com.arkivanov.mvikotlin.sample.coroutines.shared.details.DetailsView.Model
-import com.arkivanov.mvikotlin.sample.coroutines.shared.details.store.DetailsStore.Intent
-import com.arkivanov.mvikotlin.sample.coroutines.shared.details.store.DetailsStore.State
+import com.arkivanov.mvikotlin.sample.coroutines.shared.details.store.Intent
+import com.arkivanov.mvikotlin.sample.coroutines.shared.details.store.State
 
 internal val stateToModel: (State) -> Model =
     { state ->
