+++ sample/coroutines/shared/src/commonMain/kotlin/com/arkivanov/mvikotlin/sample/coroutines/shared/main/store/add/AddStoreFactory.kt
@@ -3,7 +3,7 @@ package com.arkivanov.mvikotlin.sample.coroutines.shared.main.store.add
 import com.arkivanov.mvikotlin.core.store.Store
 import com.arkivanov.mvikotlin.core.store.StoreFactory
 import com.arkivanov.mvikotlin.core.utils.JvmSerializable
-import com.arkivanov.mvikotlin.extensions.coroutines.CoroutineExecutor
+import com.arkivanov.mvikotlin.extensions.coroutines.coroutineExecutorFactory
 import com.arkivanov.mvikotlin.sample.coroutines.shared.main.store.add.AddStore.Intent
 import com.arkivanov.mvikotlin.sample.coroutines.shared.main.store.add.AddStore.Label
 import com.arkivanov.mvikotlin.sample.coroutines.shared.main.store.add.AddStore.State
@@ -13,6 +13,12 @@ import kotlinx.coroutines.launch
 import kotlinx.coroutines.withContext
 import kotlin.coroutines.CoroutineContext
 
+/**
+ * This builder function showcases the way of creating a [Store] *without*
+ * using the DSL API and *with* the dedicated interface [AddStore].
+ * This option may work better for smaller and simpler stores.
+ * The [Intent], [State] and [Label] classes are defined inside the store interface.
+ */
 internal fun StoreFactory.addStore(
     database: TodoDatabase,
     mainContext: CoroutineContext,
@@ -21,46 +27,28 @@ internal fun StoreFactory.addStore(
     object : AddStore, Store<Intent, State, Label> by create(
         name = "TodoAddStore",
         initialState = State(),
-        executorFactory = {
-            ExecutorImpl(
-                database = database,
-                mainContext = mainContext,
-                ioContext = ioContext,
-            )
+        executorFactory = coroutineExecutorFactory(mainContext) {
+            onIntent<Intent.SetText> { dispatch(Msg.TextChanged(it.text)) }
+
+            onIntent<Intent.Add> {
+                val text = state().text.takeUnless(String::isBlank) ?: return@onIntent
+
+                dispatch(Msg.TextChanged(""))
+
+                launch {
+                    val item = withContext(ioContext) { database.create(TodoItem.Data(text = text)) }
+                    publish(Label.Added(item))
+                }
+            }
         },
-        reducer = { reduce(it) },
+        reducer = { msg: Msg ->
+            when (msg) {
+                is Msg.TextChanged -> copy(text = msg.text)
+            }
+        },
     ) {}
 
 // Serializable only for exporting events in Time Travel, no need otherwise.
 private sealed class Msg : JvmSerializable {
     data class TextChanged(val text: String) : Msg()
 }
-
-private class ExecutorImpl(
-    private val database: TodoDatabase,
-    mainContext: CoroutineContext,
-    private val ioContext: CoroutineContext,
-) : CoroutineExecutor<Intent, Nothing, State, Msg, Label>(mainContext) {
-    override fun executeIntent(intent: Intent) {
-        when (intent) {
-            is Intent.SetText -> dispatch(Msg.TextChanged(intent.text))
-            is Intent.Add -> addItem()
-        }.let {}
-    }
-
-    private fun addItem() {
-        val text = state().text.takeUnless(String::isBlank) ?: return
-
-        dispatch(Msg.TextChanged(""))
-
-        scope.launch {
-            val item = withContext(ioContext) { database.create(TodoItem.Data(text = text)) }
-            publish(Label.Added(item))
-        }
-    }
-}
-
-private fun State.reduce(msg: Msg): State =
-    when (msg) {
-        is Msg.TextChanged -> copy(text = msg.text)
-    }
